Tests/LibWeb/Text/expected/async-scrolling/viewport-body-overflow-hidden-horizontal-wheel-target.txt

document scroll width: 813
horizontal wheel target: none
vertical wheel target: viewport
scroll x after horizontal wheel: 0
